+++ src/main/java/nextstep/subway/path/dto/PathResponse.java
@@ -7,12 +7,12 @@ import java.util.List;
 
 public class PathResponse {
     private List<StationResponse> stations;
-    private int distance;
+    private double distance;
 
     public PathResponse() {
     }
 
-    public PathResponse(List<StationResponse> stations, int distance) {
+    public PathResponse(List<StationResponse> stations, double distance) {
         this.stations = stations;
         this.distance = distance;
     }
@@ -21,7 +21,7 @@ public class PathResponse {
         return stations;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
 }
